Arreglos process

+++ TP2-SO-2025/Userland/Shell/commands/wc.c
@@ -8,7 +8,7 @@ int _wc(int argc, char *argv[]){
         printf("Se registro 1 linea por stdin\n"); //consideramos 1 linea los argumentos
         return 0;
     }
-    long lines = 0;
+    uint32_t lines = 0;
     int c;
     while ((c = getchar()) != EOF && c != '\0') {
         if (c == '\n') {
